bit of clean up in experimental
